Security enhancement - digitally-signed the root file system image.
The public key used to verify the signature is inside the initramfs.

This is not enough, because it's possible to create a self-signed image and re-pack the initramfs with the new public key. Signing the kernel can be done through Secure Boot, but signing the initramfs requires some clever trick.